refactor: split DaemonClient.swift into focused files
Extract connection management and message routing from the 1,719-line
DaemonClient.swift into dedicated extension files. No logic changes —
only code movement and access modifier adjustments (private → internal
for cross-file extension access).
